// frontend/src/lib/types/ProvidersFile.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ProviderConfig } from "./ProviderConfig";

/**
 * The parsed contents of a `providers.toml` file.
 */
export type ProvidersFile = { 
/**
 * Each `[[providers]]` table.
 */
providers: Array<ProviderConfig>, };